Set full node collision box for bushy leaves
When a node has no collision box it uses the node box as the collision
box. This effect made walking on nodes modified by this mod difficult.

This patch explicitly sets a collision box with the size of a full node
for modified nodes. This makes the “bushy leaves” effect purely visual;
it does not affect collision differently than if the mod is not loaded.